+++ sfsimodels/__init__.py
@@ -3,7 +3,7 @@ from sfsimodels.models.hazards import SeismicHazard
 from sfsimodels.models.foundations import Foundation, PadFoundation, RaftFoundation
 from sfsimodels.models.soils import Soil, CriticalSoil, discretize_soil_profile, SoilProfile, StressDependentSoil
 from sfsimodels.models.buildings import Building, FrameBuilding, WallBuilding, SDOFBuilding, FrameBuilding2D, \
-    NullBuilding
+    NullBuilding, BeamColumnElement
 from sfsimodels.models.sections import Section
 from sfsimodels.models import sections
 from sfsimodels.models import materials
@@ -23,6 +23,7 @@ from sfsimodels.files import ecp_dict_to_objects, load_json, loads_json, Output,
 from sfsimodels.functions import clean_float, collect_serial_value, add_to_obj, interp_left
 from sfsimodels.exceptions import DesignError, AnalysisError, ModelError, ModelWarning
 from sfsimodels import sensors
+from sfsimodels import std
 
 from sfsimodels import __about__